Removing debug traces and reorganizing code
